ui-sys: remove cmake dependency for unix
This manually compiles libui for the unix platform, following the
pattern of how the windows build was performed.

This adds a dependency on pkg_config in order to find the include paths.

Finally, a bug was encountered as part of the Unix build of bindgen.
When using clang <= 3.8, it cannot be relied on to determine whether a
symbol is C or C++.  In order to disable name mangling, we must set
`trust_clang_mangling(false)`.

Signed-off-by: Sean Cross <sean@xobs.io>